+++ Changelog.rst
@@ -6,11 +6,16 @@ version 3.3.1
 
 * New method: `cf.Field.apply_masking`
 * New method: `cf.Data.apply_masking`
-* New keyword parameters to `cf.read`: ``warn_valid``, ``mask``
+* New keyword parameter to `cf.read`: ``mask``
+* New keyword parameter to `cf.read`: ``warn_valid``
+  (https://github.com/NCAS-CMS/cfdm/issues/30)
 * New keyword parameter to `cf.write`: ``warn_valid``
+  (https://github.com/NCAS-CMS/cfdm/issues/30)
 * New keyword parameter to `cf.Field.nc_global_attributes`: ``values``
 * Added time coordinate bounds to the polygon geometry example field
   ``6`` returned by `cf.example_field`.
+* Changed exact cfdm dependency to version 1.8.3.
+* Changed minimum cfunits dependency to version 3.2.6.
 * Fixed bug in `cf.write` that caused (what are effectively)
   string-valued scalar auxiliary coordinates to not be written to disk
   as such, or even an exception to be raised.